MAINT: Rename folder for reduction to scalar data.
