+++ src/darsia/transformations/reduction/monochromatic.py
