+++ src/darsia/transformations/reduction/signalreduction.py
